Consolidate world modules
