--- world/generation.py
@@ -1,159 +1,0 @@
-"""Procedural hex map generation utilities."""
-
-from dataclasses import dataclass, field
-from enum import Enum
-from typing import Dict, Tuple
-import random
-
-# Type alias for coordinates in the hex grid
-Coordinate = Tuple[int, int]
-
-
-class ResourceType(Enum):
-    """Supported resource types found on hexes."""
-
-    WOOD = "wood"
-    STONE = "stone"
-    ORE = "ore"
-
-
-@dataclass
-class WorldSettings:
-    """Configuration values for world generation."""
-
-    seed: int = 0
-    width: int = 50
-    height: int = 50
-    biome_distribution: Dict[str, float] = field(
-        default_factory=lambda: {
-            "plains": 0.4,
-            "forest": 0.3,
-            "desert": 0.2,
-            "mountain": 0.1,
-        }
-    )
-    weather_patterns: Dict[str, float] = field(
-        default_factory=lambda: {"rain": 0.3, "dry": 0.5, "snow": 0.2}
-    )
-    moisture: float = 0.5
-    elevation: float = 0.5
-    temperature: float = 0.5
-
-
-@dataclass
-class Hex:
-    """Represents a single hex tile in the world."""
-
-    coord: Coordinate
-    biome: str = "plains"
-    elevation: float = 0.0
-    moisture: float = 0.0
-    temperature: float = 0.0
-    resources: Dict[ResourceType, int] = field(default_factory=dict)
-
-
-@dataclass
-class World:
-    """Collection of hexes generated for a world."""
-
-    settings: WorldSettings
-    hexes: Dict[Coordinate, Hex] = field(default_factory=dict)
-
-    def get_hex(self, x: int, y: int) -> Hex:
-        return self.hexes.get((x, y))
-
-    def resources_near(self, x: int, y: int, radius: int = 1) -> Dict[ResourceType, int]:
-        """Return total resources within a square radius of the coordinate."""
-        totals: Dict[ResourceType, int] = {r: 0 for r in ResourceType}
-        for dy in range(-radius, radius + 1):
-            for dx in range(-radius, radius + 1):
-                hex_ = self.get_hex(x + dx, y + dy)
-                if hex_:
-                    for rtype, amt in hex_.resources.items():
-                        totals[rtype] += amt
-        return {r: amt for r, amt in totals.items() if amt > 0}
-
-
-def initialize_random(settings: WorldSettings) -> random.Random:
-    """Create a random generator based on the provided seed."""
-    return random.Random(settings.seed)
-
-
-def generate_terrain_type(rng: random.Random, settings: WorldSettings) -> str:
-    """Choose a biome based on distribution weights."""
-    biomes = list(settings.biome_distribution.keys())
-    weights = list(settings.biome_distribution.values())
-    return rng.choices(biomes, weights=weights, k=1)[0]
-
-
-def generate_resources(rng: random.Random, biome: str) -> Dict[ResourceType, int]:
-    """Generate resources for a hex based on its biome."""
-    resources: Dict[ResourceType, int] = {}
-
-    if biome == "forest":
-        resources[ResourceType.WOOD] = rng.randint(5, 15)
-        if rng.random() < 0.3:
-            resources[ResourceType.STONE] = rng.randint(1, 4)
-    elif biome == "mountain":
-        resources[ResourceType.STONE] = rng.randint(5, 15)
-        if rng.random() < 0.7:
-            resources[ResourceType.ORE] = rng.randint(1, 5)
-    elif biome == "plains":
-        if rng.random() < 0.5:
-            resources[ResourceType.WOOD] = rng.randint(1, 5)
-        if rng.random() < 0.4:
-            resources[ResourceType.STONE] = rng.randint(1, 4)
-    elif biome == "desert":
-        if rng.random() < 0.2:
-            resources[ResourceType.STONE] = rng.randint(1, 3)
-    return resources
-
-
-def generate_hexes(settings: WorldSettings) -> World:
-    """Generate all hexes for the world according to settings."""
-    rng = initialize_random(settings)
-    world = World(settings)
-
-    for y in range(settings.height):
-        for x in range(settings.width):
-            biome = generate_terrain_type(rng, settings)
-            elevation = rng.random() * settings.elevation
-            moisture = rng.random() * settings.moisture
-            temperature = rng.random() * settings.temperature
-            resources = generate_resources(rng, biome)
-            world.hexes[(x, y)] = Hex(
-                coord=(x, y),
-                biome=biome,
-                elevation=elevation,
-                moisture=moisture,
-                temperature=temperature,
-                resources=resources,
-            )
-    return world
-
-
-def adjust_settings(
-    settings: WorldSettings,
-    *,
-    moisture: float | None = None,
-    elevation: float | None = None,
-    temperature: float | None = None,
-) -> None:
-    """Adjust world sliders before final generation."""
-    if moisture is not None:
-        settings.moisture = max(0.0, min(1.0, moisture))
-    if elevation is not None:
-        settings.elevation = max(0.0, min(1.0, elevation))
-    if temperature is not None:
-        settings.temperature = max(0.0, min(1.0, temperature))
-
-
-__all__ = [
-    "ResourceType",
-    "WorldSettings",
-    "Hex",
-    "World",
-    "generate_hexes",
-    "adjust_settings",
-]
-
